+++ internal/provider/data_source_catalystcenter_building.go
@@ -23,14 +23,10 @@ import (
 	"fmt"
 	"net/url"
 
-	"github.com/hashicorp/terraform-plugin-framework-validators/datasourcevalidator"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/path"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	cc "github.com/netascode/go-catalystcenter"
-	"github.com/tidwall/gjson"
 )
 
 // End of section. //template:end imports
@@ -63,16 +59,14 @@ func (d *BuildingDataSource) Schema(ctx context.Context, req datasource.SchemaRe
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
 				MarkdownDescription: "The id of the object",
-				Optional:            true,
-				Computed:            true,
+				Required:            true,
 			},
 			"name": schema.StringAttribute{
 				MarkdownDescription: "The name of the building",
-				Optional:            true,
 				Computed:            true,
 			},
-			"parent_name": schema.StringAttribute{
-				MarkdownDescription: "The path of the parent area, e.g. `Global/Area5`. `Global` in case of root area.",
+			"parent_id": schema.StringAttribute{
+				MarkdownDescription: "The ID of the parent area",
 				Computed:            true,
 			},
 			"country": schema.StringAttribute{
@@ -92,14 +86,6 @@ func (d *BuildingDataSource) Schema(ctx context.Context, req datasource.SchemaRe
 				Computed:            true,
 			},
 		},
-	}
-}
-func (d *BuildingDataSource) ConfigValidators(ctx context.Context) []datasource.ConfigValidator {
-	return []datasource.ConfigValidator{
-		datasourcevalidator.ExactlyOneOf(
-			path.MatchRoot("id"),
-			path.MatchRoot("name"),
-		),
 	}
 }
 
@@ -125,32 +111,10 @@ func (d *BuildingDataSource) Read(ctx context.Context, req datasource.ReadReques
 	}
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Read", config.Id.String()))
-	if config.Id.IsNull() && !config.Name.IsNull() {
-		res, err := d.client.Get("/dna/intent/api/v2/site")
-		if err != nil {
-			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve objects, got error: %s", err))
-			return
-		}
-		if value := res.Get("response"); len(value.Array()) > 0 {
-			value.ForEach(func(k, v gjson.Result) bool {
-				if config.Name.ValueString() == v.Get("name").String() {
-					config.Id = types.StringValue(v.Get("id").String())
-					tflog.Debug(ctx, fmt.Sprintf("%s: Found object with name '%v', id: %v", config.Id.String(), config.Name.ValueString(), config.Id.String()))
-					return false
-				}
-				return true
-			})
-		}
-
-		if config.Id.IsNull() {
-			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to find object with name: %s", config.Name.ValueString()))
-			return
-		}
-	}
 
 	params := ""
-	params += "?id=" + url.QueryEscape(config.Id.ValueString())
-	res, err := d.client.Get("/dna/intent/api/v2/site" + params)
+	params += "/" + url.QueryEscape(config.Id.ValueString())
+	res, err := d.client.Get(config.getPath() + params)
 	if err != nil {
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object, got error: %s", err))
 		return